Add option to hide taskbar icon when window closes
New setting 'Hide taskbar icon when Ditto window closes' in the advanced general options.
“Show in taskbar” is enabled firstly.
Closing Ditto with the “X” button or the Escape key will hide the taskbar icon. So the taskbar icon is visible only when the quick paste window is visible (similar behavior to keepass window).

Implements corresponding getter and setter in CGetSetOptions, and updates CQPasteWnd to respect this setting when hiding the main window.